GS: Move OpenGL helpers from common to GS
