+++ pcsx2/GS/Renderers/OpenGL/GLContext.cpp
@@ -1,0 +1,92 @@
+/*  PCSX2 - PS2 Emulator for PCs
+ *  Copyright (C) 2002-2021  PCSX2 Dev Team
+ *
+ *  PCSX2 is free software: you can redistribute it and/or modify it under the terms
+ *  of the GNU Lesser General Public License as published by the Free Software Found-
+ *  ation, either version 3 of the License, or (at your option) any later version.
+ *
+ *  PCSX2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *  PURPOSE.  See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with PCSX2.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "PrecompiledHeader.h"
+
+#include "GS/Renderers/OpenGL/GLContext.h"
+
+#if defined(_WIN32)
+#include "GS/Renderers/OpenGL/GLContextWGL.h"
+#elif defined(__APPLE__)
+#include "GS/Renderers/OpenGL/GLContextAGL.h"
+#else // Linux
+#ifdef X11_API
+#include "GS/Renderers/OpenGL/GLContextEGLX11.h"
+#endif
+#ifdef WAYLAND_API
+#include "GS/Renderers/OpenGL/GLContextEGLWayland.h"
+#endif
+#endif
+
+#include "common/Console.h"
+
+#include "glad.h"
+
+GLContext::GLContext(const WindowInfo& wi)
+	: m_wi(wi)
+{
+}
+
+GLContext::~GLContext() = default;
+
+std::unique_ptr<GLContext> GLContext::Create(const WindowInfo& wi)
+{
+	// We need at least GL3.3.
+	static constexpr Version vlist[] = {
+		{4, 6},
+		{4, 5},
+		{4, 4},
+		{4, 3},
+		{4, 2},
+		{4, 1},
+		{4, 0},
+		{3, 3},
+	};
+
+	std::unique_ptr<GLContext> context;
+#if defined(_WIN32)
+	context = GLContextWGL::Create(wi, vlist);
+#elif defined(__APPLE__)
+	context = GLContextAGL::Create(wi, vlist);
+#else // Linux
+#if defined(X11_API)
+	if (wi.type == WindowInfo::Type::X11)
+		context = GLContextEGLX11::Create(wi, vlist);
+#endif
+
+#if defined(WAYLAND_API)
+	if (wi.type == WindowInfo::Type::Wayland)
+		context = GLContextEGLWayland::Create(wi, vlist);
+#endif
+#endif
+
+	if (!context)
+		return nullptr;
+
+	// NOTE: Not thread-safe. But this is okay, since we're not going to be creating more than one context at a time.
+	static GLContext* context_being_created;
+	context_being_created = context.get();
+
+	// load up glad
+	if (!gladLoadGLLoader([](const char* name) { return context_being_created->GetProcAddress(name); }))
+	{
+		Console.Error("Failed to load GL functions for GLAD");
+		return nullptr;
+	}
+
+	context_being_created = nullptr;
+
+	return context;
+}
