+++ pcsx2/GS/Renderers/OpenGL/GLState.h
@@ -18,6 +18,8 @@
 #include "GS/Renderers/OpenGL/GLLoader.h"
 #include "GS/GSVector.h"
 
+#include "glad.h"
+
 class GSTextureOGL;
 
 namespace GLState
